- Moved LOVs to the Params Controller. - Renamed the AG Controller to FeesGrid (to match the partial). - Renamed the Home view as FeesByType.
