+++ SeaWideExpress/QuotingTool/AJC.Logistics.SeaWideExpress.QuotingTool.Controllers/AJC.Logistics.SeaWideExpress.QuotingTool.Controllers.csproj
@@ -114,8 +114,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ErrorController.cs" />
-    <Compile Include="AGController.cs" />
     <Compile Include="FeesByTypeController.cs" />
+    <Compile Include="ParamsController.cs" />
+    <Compile Include="FeesGridController.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
